pyproject.toml
[project]
name = "gemini-docs-mcp"
version = "0.1.0"
description = "MCP server that uses Gemini as RAG for documentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.6.1",
    "httpx>=0.28.1",
    "litellm>=1.80.10",
    "mcp>=1.24.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.12",
    "starlette>=0.38.0",
]

[project.scripts]
gemini-docs-mcp = "src.gemini_docs_mcp.server:main"

[dependency-groups]
dev = []
